Finalizing functionality
Game works as intended with the necessary rules set. 
JavaScript is logically complete allowing for a full game of tic-tac-toe to be played until a winner is determined. 
After declaring a winner the game is no longer running and does not start up again until the user restarts with a button.